Remplacé les types 'long' par des 'int' pour les egids et les numéros Ofs.
git-svn-id: https://svn.etat-de-vaud.ch/SVN_SOCLE/registre/unireg/trunk@55711 cb3efa0a-b8b0-9f43-a4c8-343d8b7e4362

+++ unireg/business/src/main/java/ch/vd/uniregctb/interfaces/model/CommuneId.java
@@ -11,7 +11,7 @@ public interface CommuneId {
 	/**
 	 * @return le numéro Ofs de la commune.
 	 */
-	long getNoOfs();
+	int getNoOfs();
 
 	/**
 	 * @return le numéro technique de la commune (uniquement nécessaire pour distinguer les communes de Lussery et Lussery-Villars qui partagent le même numéro Ofs = 5487).
